exec: add sorter
```
BenchmarkSort/rows=2048/cols=1-8                    5000            287366 ns/op          57.01 MB/s
BenchmarkSort/rows=2048/cols=2-8                    5000            292186 ns/op         112.15 MB/s
BenchmarkSort/rows=2048/cols=4-8                    5000            318730 ns/op         205.62 MB/s
BenchmarkSort/rows=16384/cols=1-8                   1000           2270545 ns/op          57.73 MB/s
BenchmarkSort/rows=16384/cols=2-8                    500           2425301 ns/op         108.09 MB/s
BenchmarkSort/rows=16384/cols=4-8                    500           2692579 ns/op         194.72 MB/s
BenchmarkSort/rows=262144/cols=1-8                    30          34905432 ns/op          60.08 MB/s
BenchmarkSort/rows=262144/cols=2-8                    30          38191955 ns/op         109.82 MB/s
BenchmarkSort/rows=262144/cols=4-8                    30          47341978 ns/op         177.19 MB/s
```

It appears we're fully bound on Go's sort implementation. I suspect that
removing the n log n interface lookups would speed this up a bunch. It's
annoying that Go doesn't have templated sort.


Release note: None

+++ pkg/sql/exec/execgen/cmd/execgen/colvec_gen.go
@@ -64,7 +64,7 @@ func (m *memColumn) AppendWithSel(
 	}
 }
 
-func (m *memColumn) Copy(src ColVec, srcStartIdx, srcEndIdx int, typ types.T) {
+func (m *memColumn) Copy(src ColVec, srcStartIdx, srcEndIdx uint64, typ types.T) {
 	switch typ {
 	{{range .}}
 	case types.{{.ExecType}}:
